Add project to check C++11 support using CMake
Many C++ projects are migrating to the new C++11 standard (and the soon
to be available C++14 standard). Compilers and standard libraries,
whilst gradually stabilizing, are still in a state of flux regarding
which features are supported. There is therefore a need to check the
compiler and library at configuration time to determine which features
can be used.

Start a new project to investigate the use of CMake and its
configuration/test system to determine available C++YY features.
Use the cmake-cxx11 modules from

git://anongit.kde.org/scratch/dakon/cmake-cxx11.git

(viewable at
http://quickgit.kde.org/?p=scratch%2Fdakon%2Fcmake-cxx11.git)

as a starting point.

Add top level CMake script to start project and include the main
CXX checking module, which then automatically performs the checks.