Update description of project and usage
Project is currently in a grey area - whilst it can be "cmaked" this
is for nothing more than demonstration and testing. It is also
not yet in a true "cmake module" state because of the top level
CMake script.

Note this in the README, with two subsections in README to describe
these different use cases.