Document issues with checks using Xcode generator
On OS X with Clang and Makefiles, it is sufficient the choice of
standard library can be made by adding the -stdlib argument to
CMAKE_CXX_FLAGS. When using the Xcode generator in the same way,
checks of C++11 features fail. This was diagnosed as

- Xcode generator does not forward CMAKE_CXX_FLAGS to the linker,
resulting in Xcode's default C++ library being used.
- "Older" Xcode versions will default to linking to the GNU library.

The breakpoint in "Older" has not yet been identified.

Document these issues in README and feature checking modules. Add a
check for support of the -stdlib argument by the compiler as a
preliminary step in identifying the standard library settings that
try_run and try_compile may need. Add, but comment out, CMAKE_FLAGS
arguments to use of try_run/compile calls, setting the Xcode attribute
required to enable libc++.

TODO: Test on Mavericks with newer Xcode, confirm that setting Xcode
attributes does not affect Makefile or other generators.